use lazy_buffer in combinations_with_replacement
